Improved PGO
Benchmarks for all selected SIMD instruction sets are now run
during profiling. If not all of them are supported, the build
script doesn't do PGO at all. I do it that way because otherwise
GCC thinks that the code that was not run in the benchmark is
never run, and generates small, but slow code for it.