Changed the semantics of FFT.alignment and pfft_alignment_f.
FFT.alignment now returns the minimal alignement needed for other
functions to work. Aligning to any greater power of two for
performance reasons should be done by the user, if that is desired.

Also changed the functions that allocate in pfft.clib to align to
something greater than what FFT.align returns.